pyproject.toml
[tool.poetry]
name = "pcl-pose-estimation"
version = "0.1.0"
description = ""
authors = ["Yarden <yarden.as@inf.ethz.ch>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10,<3.13"
equinox = "^0.11.1"
jaxlib = {version = "^0.4.2+cuda12.cudnn82"}
optax = "^0.1.7"
hydra-core = "^1.3.2"


[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[tool.poetry.group.dev.dependencies]
mypy = "^1.7.0"
pre-commit = "^3.5.0"
ruff = "^0.1.5"
pytest = "^7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
plugins = ["numpy.typing.mypy_plugin"]
ignore_missing_imports = true
show_column_numbers = true
allow_untyped_defs = true
allow_subclassing_any = true
allow_untyped_calls = true
strict_optional = true
